long long
 - changed to `long long`, almost always 64bit as opposed to just `long`.

+++ kj.h
@@ -206,8 +206,8 @@ typedef signed short i16;
 typedef unsigned short u16;
 typedef signed int i32;
 typedef unsigned int u32;
-typedef signed long i64;
-typedef unsigned long u64;
+typedef signed long long i64;
+typedef unsigned long long u64;
 #endif
 
 #if !defined(I8_MIN)
